backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.13
python-dotenv==1.0.0
openai==1.3.0
chromadb==0.4.21
langgraph==0.0.32
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.28
httpx==0.27.0
requests==2.31.0
duckduckgo-search==6.2.13
networkx==3.2.1
rank_bm25==0.2.2
uvloop==0.19.0
python-json-logger==2.0.7
tenacity==8.2.3
prometheus_client==0.19.0
python-multipart==0.0.6
